// projectThree/project-3-front-end-beta-main/project-3-front-end-beta-main/src/app/models/user.model.ts
export class User {

    userId: number = 0;
    userEmail: string = "";
    userName: string = "";
    userPassword: string = "";
    userFirstName: string = "";
    userLastName: string = "";
    userAddress: string = "";
    userType: string = "";
    userContact: string = "";
    userRemoved: boolean = false;
}